% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_stat}
\alias{bootstrap_stat}
\title{Bootstrap statistic estimator}
\usage{
bootstrap_stat(x, stat = mean, B = 1000, n = NULL, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{stat}{Function used to compute the statistic. Default is mean.}

\item{B}{Number of bootstrap samples. Default is 1000.}

\item{n}{Size of each bootstrap sample. If NULL (default), uses the
number of non-missing observations in \code{x}.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector containing B bootstrap statistic estimates.
}
\description{
Generate bootstrap estimates of a statistic using resampling
with replacement.
}
