% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{draw_histogram}
\alias{draw_histogram}
\title{Create a histogram with the inputted values}
\usage{
draw_histogram(
  xs,
  draw_lines = TRUE,
  dist,
  a = 0,
  b = 0,
  plot_color = "skyblue",
  plot_border = "white",
  line_color = "red",
  line_width = 2
)
}
\arguments{
\item{xs}{a vector of simulated values.}

\item{draw_lines}{draws a line to represent the pdf of the chosen distribution, TRUE by default.}

\item{dist}{optional - sets the distribution being simulated to generate its actual pdf, accepts "kumaraswamy" and "arcsine".}

\item{a}{optional - sets the first parameter of the Kumaraswamy distribution.}

\item{b}{optional - sets the other parameter of the Kumaraswamy distribution.}

\item{plot_color}{sets the color of the bars of the histogram.}

\item{plot_border}{sets the color of the borders of the bars.}

\item{line_color}{sets the color of the pdf line.}

\item{line_width}{sets the width of the line.}
}
\description{
Create a histogram with an optional line to represent the theoretical pdf of the Kumaraswamy distribution or the arcsine distribution.
}
\examples{
set.seed(1)
n <- 50000
a <- 0.5
b <- 0.5
xs <- is_kumaraswamy(n, a, b)
draw_histogram(xs, dist="kumaraswamy", a=a, b=b)
}
