% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{draw_index_scatter}
\alias{draw_index_scatter}
\title{Create a scatterplot with the inputted values}
\usage{
draw_index_scatter(xs, point_color = "black", point_size = 0.6, alpha = 0.6)
}
\arguments{
\item{xs}{a vector of simulated values.}

\item{point_color}{sets the color of the data points on the scatterplot.}

\item{point_size}{sets the size of the data points on the scatterplot.}

\item{alpha}{sets the transparency (alpha) of the data points on the scatterplot.}
}
\description{
Create a scatterplot to display the values in a vector. Can be used for the simulated values of Kumaraswamy distribution or the arcsine distribution.
}
\examples{
set.seed(1)
n <- 50000
a <- 0.5
b <- 0.5
xs <- is_kumaraswamy(n, a, b)
draw_index_scatter(xs, point_color="black", point_size=0.6, alpha=0.6)
}
