% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{draw_table}
\alias{draw_table}
\title{Draw a table of statistics for comparison}
\usage{
draw_table(xs, dist, a = 0, b = 0)
}
\arguments{
\item{xs}{a vector of simulated values.}

\item{dist}{sets the distribution being simulated to generate its theoretical values, accepts "kumaraswamy" and "arcsine".}

\item{a}{optional - sets the first parameter of the Kumaraswamy distribution.}

\item{b}{optional - sets the other parameter of the Kumaraswamy distribution.}
}
\description{
Create a table comparing the simulated statistics with the theoretical statistics.
}
\examples{
set.seed(1)
n <- 50000
a <- 0.5
b <- 0.5
xs <- is_kumaraswamy(n, a, b)
draw_table(xs, "kumaraswamy", a, b)
}
