% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{is_arcsine}
\alias{is_arcsine}
\title{Arcsine distribution simulation through inverse transformation}
\usage{
is_arcsine(n)
}
\arguments{
\item{n}{sets the numbers of samples to be generated.}
}
\value{
An nx1 matrix of the values to simulate the arcsine distribution.
}
\description{
Arcsine distribution simulation through inverse transformation
}
\examples{
set.seed(1)
n <- 50000
xs <- is_arcsine(n)
head(xs)
}
