% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{is_kumaraswamy}
\alias{is_kumaraswamy}
\title{Kumaraswamy distribution simulation through inverse transformation}
\usage{
is_kumaraswamy(n, a, b)
}
\arguments{
\item{n}{sets the number of samples to be generated.}

\item{a}{sets the first parameter of the distribution.}

\item{b}{sets the other parameter of the distribution.}
}
\value{
An nx1 matrix of the values to simulate the Kumaraswamy distribution.
}
\description{
Kumaraswamy distribution simulation through inverse transformation
}
\examples{
set.seed(1)
n <- 50000
a <- 0.5
b <- 0.5
xs <- is_kumaraswamy(n, a, b)
head(xs)
}
