% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransformation.R
\name{kumaraswamy_moment}
\alias{kumaraswamy_moment}
\title{Helper function generating moments of the Kumaraswamy distribution}
\usage{
kumaraswamy_moment(n, a, b)
}
\arguments{
\item{n}{sets the nth moment of the distribution.}

\item{a}{sets the first parameter of the distribution.}

\item{b}{sets the other parameter of the distribution.}
}
\value{
the nth moment of Kumaraswamy(a, b)
}
\description{
Helper function generating moments of the Kumaraswamy distribution
}
