% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_mean}
\alias{bootstrap_mean}
\title{Bootstrap mean estimator}
\usage{
bootstrap_mean(x, B = 1000, n = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{B}{Number of bootstrap samples. Default is 1000.}

\item{n}{Size of each bootstrap sample. If NULL (default), uses the
number of non-missing observations in \code{x}.}
}
\value{
A numeric vector containing B bootstrap mean estimates.
}
\description{
Generate bootstrap estimates of the sample mean using resampling
with replacement.
}
