% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_median}
\alias{bootstrap_median}
\title{Bootstrap median estimator}
\usage{
bootstrap_median(x, B = 1000, n = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{B}{Number of bootstrap samples. Default is 1000.}

\item{n}{Size of each bootstrap sample. If NULL (default), uses the
number of non-missing observations in \code{x}.}
}
\value{
Numeric vector containing B bootstrap median estimates.
}
\description{
Generate bootstrap estimates of the sample median using resampling
with replacement.
}
