% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compute_sleep_duration}
\alias{compute_sleep_duration}
\title{Compute sleep duration}
\usage{
compute_sleep_duration(bedtime, uptime)
}
\arguments{
\item{bedtime}{Numeric vector of bedtimes.}

\item{uptime}{Numeric vector of wake times.}
}
\value{
A numeric vector containing sleep duration (in hours).
}
\description{
Compute sleep duration as the difference between wake time and bedtime.
If wake time is smaller than bedtime, the function assumes sleep crosses
midnight and adjusts accordingly.
}
